Fix scroll bug, reimplement metadata
A bug was introduced in commit d0fa7ab in which scroll positions were
not reset when switching to a new page.
This was caused by the placement of the `PageMeta` component at the top
of the page.
This has been fixed by replacing `PageMeta` with an exported `metadata`
object which is the proper way of adding metadata in Next.js.
As a positive side effect, there is now clearer separation between
Client Components and Server Components.

Additions, in `/src/app`:
* `components/misc.js` for miscellaneous helper functions
* `creators/contact/CreatorForm.jsx` for the creator-form (see below)
* `about/client-components.jsx` for the about page (see below)
* `faq/Content.jsx` for the FAQ page (see below)
* `contact/Content.jsx` for the contact page (see below)

Resolved warning about the spelling and type of `webkit-playsinline` in:
* Carousel component
* Home 1st
* Home `VideoBig` component
* Brands 1st
* Creators 1st
* Creators 2nd

Favicon:
* Moved and renamed from `/public/test1.png` to `/src/app/icon.png`.
  This is the proper Next.js way of adding a favicon.

OpenGraph image:
* Moved and renamed from `/public/images/og-image.png` to
  `/src/app/opengraph-image.png`.
  This is the proper Next.js way of adding an OpenGraph image.

`PageMeta` component:
* Removed metadata-setting functionality.
  It will be removed entirely once we've migrated to the official
  Next.js way of enabling Google Analytics.

`Companies` component:
* Resolved warning about list children lacking a `key` prop

Root layout:
* Changed `metadata` to serve as a template for other pages

Brands-form component:
* Added `"use client"`.
  This fixes an error caused by removing it from several pages.

Home page:
* Replaced `PageMeta` with `metadata`
  * Removed `"use client"`.
    `metadata` cannot be exported when it is declared.

Home `VideoSmall` component:
* Minor cleanups

Brands page:
* Replaced `PageMeta` with `metadata`
  * Removed `"use client"` (see brands 4th and last)

Brands 4th:
* Added `"use client"`.
  This fixes an error caused by removing it from the brands page.

Brands last:
* Added `"use client"`.
  This fixes an error caused by removing `"use client"` from several
  pages.

Creators page:
* Replaced `PageMeta` with `metadata`
  * Removed `"use client"` (see brands last)
* Implemented lazy loading

Creator-contact, about, FAQ, contact pages:
* Replaced `PageMeta` with `metadata`
  * Moved Client Components to another file

Resources, newsroom, privacy-policy, terms-and-conditions pages:
* Replaced `PageMeta` with `metadata`
  * Removed unnecessary `"use client"`

Resource-content module:
* Simplified method of adding keywords
  * Simplified overall page structure

Resource articles 1 to 6:
* Replaced `PageMeta` with `metadata`
  * Removed unnecessary `"use client"`
* Added meta keywords
