+++ src/app/components/Carousel.jsx
@@ -26,7 +26,7 @@ const Carousel = ({
             autoPlay
             loop
             muted
-            webkit-playsInline
+            webkit-playsinline="true"
             playsInline
             preload="metadata"
           >
