+++ src/app/components/Form.jsx
@@ -1,3 +1,4 @@
+"use client";
 import Image from "next/image";
 import { useState } from "react";
 import { motion } from "framer-motion";
