+++ src/app/components/PageMeta.jsx
@@ -1,29 +1,28 @@
-// Component for page metadata and initializing Google Analytics
-
 import ReactGA from "react-ga4";
 
 export const PageMeta = ({
-  title,
-  desc,
-  // The path portion of the page's canonical URL
-  // Example: if the canonical URL is `http://www.example.com/example/page.html`, then `canonicalPath` is `/example/page.html`
-  canonicalPath,
+  // title,
+  // desc,
+  // // The path portion of the page's canonical URL
+  // // Example: if the canonical URL is `http://www.example.com/example/page.html`, then `canonicalPath` is `/example/page.html`
+  // canonicalPath,
 
-  // OpenGraph properties: https://ogp.me/
-  // Guidelines: https://ahrefs.com/blog/open-graph-meta-tags/
-  //
-  // If unspecified, equal to `title`
-  ogTitle,
-  // If unspecified, defaults to `"website"`
-  ogType,
-  // If unspecified, defaults to 1200x630 Cult Creative logo
-  ogImage,
-  // If unspecified, equal to `desc`
-  ogDesc
+  // // OpenGraph properties: https://ogp.me/
+  // // Guidelines: https://ahrefs.com/blog/open-graph-meta-tags/
+  // //
+  // // If unspecified, equal to `title`
+  // ogTitle,
+  // // If unspecified, defaults to `"website"`
+  // ogType,
+  // // If unspecified, defaults to 1200x630 Cult Creative logo
+  // ogImage,
+  // // If unspecified, equal to `desc`
+  // ogDesc
 }) => {
+  // Comment everything out except for this, we're going to switch to the Next.js way of adding metadata
   ReactGA.initialize("G-NP1X4Y7S7R");
 
-  const canonicalUrl = `https://www.cultcreativeasia.com${canonicalPath}`;
+  // const canonicalUrl = `https://www.cultcreativeasia.com${canonicalPath}`;
 
   return (
     <>
@@ -33,17 +32,17 @@ export const PageMeta = ({
       Apparently, using a hyphen instead of a pipe as a separator improves SEO:
       https://www.semrush.com/blog/case-study-should-you-add-pipes-or-dashes-to-your-title-ag-/
       */}
-      <title>{`${title} - Cult Creative`}</title>
+      {/* <title>{`${title} - Cult Creative`}</title>
       <meta name="description" content={desc} />
-      <link rel="canonical" href={canonicalUrl} />
+      <link rel="canonical" href={canonicalUrl} /> */}
 
       {/* Do not insert `<title>` suffix; guidelines recommend that branding be left out */}
-      <meta property="og:title" content={ogTitle || title} />
+      {/* <meta property="og:title" content={ogTitle || title} />
       <meta property="og:type" content={ogType || "website"} />
       <meta property="og:image" content={ogImage || "/images/og-image.png"} />
-      <meta property="og:url" content={canonicalUrl} />
+      <meta property="og:url" content={canonicalUrl} /> */}
       {/* This property is optional, but `desc` is a nice default to have */}
-      <meta property="og:description" content={ogDesc || desc} />
+      {/* <meta property="og:description" content={ogDesc || desc} /> */}
     </>
   );
 };
