+++ src/app/layout.js
@@ -4,20 +4,25 @@ import "./globals.css";
 const inter = Inter({ subsets: ["latin"] });
 
 export const metadata = {
-  title: "Cult Creative",
-  description: "Quality UGC creators for your marketing, at scale.",
+  title: {
+    // Apparently, using a hyphen instead of a pipe as a separator improves SEO:
+    // https://www.semrush.com/blog/case-study-should-you-add-pipes-or-dashes-to-your-title-ag-/
+    template: "%s - Cult Creative",
+    default: "Cult Creative",
+  },
+  // TODO BUG: Rightfully, `/src/app/opengraph-image.png` should be the default for all pages, but somehow it only applies to the home page
+  // Uncomment this code if this bug persists in the final website
+  // openGraph: {
+  //   images: [ { url: "/opengraph-image.png" }]
+  // },
 };
 
 const RootLayout = ({ children }) => {
   return (
     <html lang="en">
-      <link
-        rel="icon"
-        // href="https://static.wixstatic.com/media/32f93d_e99c7f1506164a29a738412bee764493%7Emv2.png/v1/fill/w_192%2Ch_192%2Clg_1%2Cusm_0.66_1.00_0.01/32f93d_e99c7f1506164a29a738412bee764493%7Emv2.png"
-        href="/test1.png"
-        type="image/png"
-      />
-      <body className={inter.className}>{children}</body>
+      <body className={inter.className}>
+        {children}
+      </body>
     </html>
   );
 };
