+++ src/app/page.js
@@ -1,5 +1,5 @@
-"use client";
 import dynamic from "next/dynamic";
+import { makeCanonicalUrl } from "@/app/components/misc";
 import { PageMeta } from "@/app/components/PageMeta";
 import Header from "@/app/components/Header";
 import { Suspense } from "react";
@@ -17,30 +17,36 @@ const HomeFive = dynamic(() => import("@/app/sections/homeSection/homeFive"));
 const HomeSix = dynamic(() => import("@/app/sections/homeSection/homeSix"));
 const LastSection = dynamic(() => import("@/app/sections/brandsection/LastSection"));
 
+export const metadata = {
+  title: {
+    absolute: "Cult Creative",
+  },
+  description: "Find top creators from across Southeast Asia for your next project. Trusted by leading companies, our platform simplifies the hiring process for both parties.",
+  openGraph: {
+    title: "Discover UGC Creators, Brand Deals and Creator Marketing Solutions with SEA’s Chosen Platform",
+  },
+  alternates: {
+    canonical: makeCanonicalUrl("/"),
+  },
+};
+
 const Home = () => {
   return (
-    <>
-      <PageMeta
-        title="Welcome"
-        desc="Find top creators from across Southeast Asia for your next project. Trusted by leading companies, our platform simplifies the hiring process for both parties."
-        canonicalPath="/"
-        ogTitle="Discover UGC Creators, Brand Deals and Creator Marketing Solutions with SEA’s Chosen Platform"
-      />
-      <main className="flex min-h-screen flex-col max-w-[100vw]">
-        <Header />
-        <Suspense fallback={<p>Loading...</p>}>
-          <HomeFirst />
-          <HomeTwo />
-          <HomeThree />
-          <HomeCompanies title="trusted by 900+ companies" />
-          <HomeFour />
-          <HomeFive />
-          <HomeSix />
-          <LastSection color="#1340FF" />
-          <Toaster position="top-center" />
-        </Suspense>
-      </main>
-    </>
+    <main className="flex min-h-screen flex-col max-w-[100vw]">
+      <PageMeta />
+      <Header />
+      <Suspense fallback={<p>Loading...</p>}>
+        <HomeFirst />
+        <HomeTwo />
+        <HomeThree />
+        <HomeCompanies title="trusted by 900+ companies" />
+        <HomeFour />
+        <HomeFive />
+        <HomeSix />
+        <LastSection color="#1340FF" />
+        <Toaster position="top-center" />
+      </Suspense>
+    </main>
   );
 };
