+++ src/app/resources/key-ugc-content-creation-ideas-and-how-to-stay-inspired/page.jsx
@@ -1,145 +1,151 @@
-"use client";
-import { PageMeta } from "@/app/components/PageMeta";
 import resources from "@/app/contants/resources.json";
+import { makeCanonicalUrl } from "@/app/components/misc";
+import { PageMeta } from "@/app/components/PageMeta";
 import * as rc from "@/app/components/resource-content";
 
-const Resource4 = () => {
-  const title = "Key UGC Content Creation Ideas and How to Stay Inspired";
+const title = "Key UGC Content Creation Ideas and How to Stay Inspired";
 
+const keywords = [
+  "Content creation inspiration",
+  "Motivation for content creators",
+  "Quality content strategies",
+  "Creative community engagement",
+  // Use a normal single quote here because users are not likely to type out the proper apostrophe character
+  "Overcoming creator's block",
+  "Creator economy in Malaysia"
+];
+
+export const metadata = {
+  title: title,
+  description: "Stay motivated by setting goals and engaging with a supportive community. Create high-quality content by understanding your audience, prioritising consistency and using engaging visuals.",
+  keywords: keywords,
+  openGraph: {
+    images: [{ url: resources[3].image }],
+  },
+  alternates: {
+    canonical: makeCanonicalUrl("/resources/key-ugc-content-creation-ideas-and-how-to-stay-inspired"),
+  },
+};
+
+const Resource4 = () => {
   return (
     <>
-      <PageMeta
-        title={title}
-        desc="Stay motivated by setting goals and engaging with a supportive community. Create high-quality content by understanding your audience, prioritising consistency and using engaging visuals."
-        canonicalPath="/resources/key-ugc-content-creation-ideas-and-how-to-stay-inspired"
-        ogImage={resources[3].image}
-      />
-      <rc.Page>
-        <rc.Content>
-          <rc.h1>{title}</rc.h1>
+      <PageMeta />
+      <rc.Page keywords={keywords}>
+        <rc.h1>{title}</rc.h1>
 
-          <rc.h2>Staying Motivated and Inspired as a Content Creator</rc.h2>
-          <rc.p>
-            Let’s be real, sometimes you just don’t feel like creating content,
-            especially when you hit that dreary creator’s block. But when you’re
-            working with clients or even if you’re just looking to build a loyal
-            fanbase, you need to stay motivated to create content.
-          </rc.p>
-          <rc.p>
-            One way to do this is by setting specific goals for your content
-            creation, whether it’s increasing your audience, collaborating with
-            other creators, or improving your skills in a particular area.
-          </rc.p>
-          <rc.p>
-            Surrounding yourself with a supportive and creative community can also
-            help inspire new ideas and motivate you to keep creating. Further,
-            engaging with your audience through comments, messages, and feedback
-            can also provide you with valuable insights and encouragement.
-          </rc.p>
+        <rc.h2>Staying Motivated and Inspired as a Content Creator</rc.h2>
+        <rc.p>
+          Let’s be real, sometimes you just don’t feel like creating content,
+          especially when you hit that dreary creator’s block. But when you’re
+          working with clients or even if you’re just looking to build a loyal
+          fanbase, you need to stay motivated to create content.
+        </rc.p>
+        <rc.p>
+          One way to do this is by setting specific goals for your content
+          creation, whether it’s increasing your audience, collaborating with
+          other creators, or improving your skills in a particular area.
+        </rc.p>
+        <rc.p>
+          Surrounding yourself with a supportive and creative community can also
+          help inspire new ideas and motivate you to keep creating. Further,
+          engaging with your audience through comments, messages, and feedback
+          can also provide you with valuable insights and encouragement.
+        </rc.p>
 
-          <rc.h2>Where to Look for Content Creation Ideas</rc.h2>
-          <rc.p>
-            When searching for content creation ideas, it can be helpful to draw
-            inspiration from various sources. Some potential sources for content
-            creation ideas include:
-          </rc.p>
-          <rc.ul>
-            <rc.li>
-              <b>Exploring trending topics</b> Staying on top of trending topics
-              and current events can provide a wealth of inspiration for content
-              creation. From popular challenges to viral news stories, there’s
-              always something new to explore and centre your content around.
-            </rc.li>
-            <rc.li>
-              <b>Collaborations and networking.</b> Collaborating with other
-              creators and industry professionals can open the door to fresh ideas
-              and innovative content formats, bringing new perspectives to your
-              content creation process.
-            </rc.li>
-            <rc.li>
-              <b>Personal experience and reflection.</b> Drawing from your own
-              experiences and personal reflections can lead to authentic and
-              relatable content. Sharing stories, lessons, or behind-the-scenes
-              glimpses into your creative process can resonate with your audience
-              and foster a deeper connection too.
-            </rc.li>
-            <rc.li>
-              <b>Industry influencers and thought leaders.</b> Following industry
-              influencers and thought leaders in your field may provide
-              inspiration for your next content idea as you work to expand on
-              their stories and content.
-            </rc.li>
-          </rc.ul>
+        <rc.h2>Where to Look for Content Creation Ideas</rc.h2>
+        <rc.p>
+          When searching for content creation ideas, it can be helpful to draw
+          inspiration from various sources. Some potential sources for content
+          creation ideas include:
+        </rc.p>
+        <rc.ul>
+          <rc.li>
+            <b>Exploring trending topics</b> Staying on top of trending topics
+            and current events can provide a wealth of inspiration for content
+            creation. From popular challenges to viral news stories, there’s
+            always something new to explore and centre your content around.
+          </rc.li>
+          <rc.li>
+            <b>Collaborations and networking.</b> Collaborating with other
+            creators and industry professionals can open the door to fresh ideas
+            and innovative content formats, bringing new perspectives to your
+            content creation process.
+          </rc.li>
+          <rc.li>
+            <b>Personal experience and reflection.</b> Drawing from your own
+            experiences and personal reflections can lead to authentic and
+            relatable content. Sharing stories, lessons, or behind-the-scenes
+            glimpses into your creative process can resonate with your audience
+            and foster a deeper connection too.
+          </rc.li>
+          <rc.li>
+            <b>Industry influencers and thought leaders.</b> Following industry
+            influencers and thought leaders in your field may provide
+            inspiration for your next content idea as you work to expand on
+            their stories and content.
+          </rc.li>
+        </rc.ul>
 
-          <rc.h2>How Do I Make Good Quality Content?</rc.h2>
-          <rc.p>
-            Creating good quality content involves a combination of factors that
-            contribute to its overall appeal and value. Here are some key
-            strategies to consider when striving to produce high-quality content:
-          </rc.p>
-          <rc.ul>
-            <rc.li>
-              <b>Understanding your audience.</b> One of the fundamental aspects
-              of creating good quality content is to understand your audience.
-              Conduct thorough research to determine the demographics,
-              preferences, and interests of your target audience. This allows you
-              to tailor your content so that it resonates with them, at once
-              increasing engagement and relevance.
-            </rc.li>
-            <rc.li>
-              <b>Consistency is key.</b> Consistently creating and publishing
-              content is essential for maintaining a strong presence and
-              connection with your audience. Whether it’s a weekly blog post,
-              daily social media updates, or regular video uploads, maintaining a
-              consistent schedule helps build anticipation and reliability among
-              your audience.
-            </rc.li>
-            <rc.li>
-              <b>Quality over quantity.</b> While consistency is important, it’s
-              equally crucial to prioritise quality over quantity. Focus on
-              producing content that is well-researched, valuable, and relevant to
-              your audience’s needs and interests. Strive for originality and
-              uniqueness in your content to stand out in a crowded digital
-              landscape.
-            </rc.li>
-            <rc.li>
-              <b>Use engaging visual elements.</b> Incorporating engaging visuals
-              such as images, infographics, and videos can significantly enhance
-              the overall appeal of your content. Visual elements not only capture
-              attention, but also help in conveying your messages effectively.
-            </rc.li>
-            <rc.li>
-              <b>Storytelling and emotion.</b> Effective storytelling can evoke
-              powerful emotions and create a lasting impact on your audience.
-              Consider incorporating compelling narratives, personal anecdotes,
-              and emotional appeal in your content to establish a deeper
-              connection with your audience.
-            </rc.li>
-            <rc.li>
-              <b>Continuous improvement.</b> Always seek opportunities for
-              improvement. Stay updated with the latest trends, gather feedback
-              from your audience, and be open to learn and evolve as a content
-              creator. Embracing a mindset of continuous improvement ensures that
-              your content remains relevant and compelling in an ever-changing
-              digital landscape.
-            </rc.li>
-          </rc.ul>
-          <rc.p>
-            While it is the goal to constantly create content, it is also crucial
-            to take breaks when needed as burnout can really impact your
-            creativity. Finding a balance between work and personal life is
-            essential for long-term motivation and inspiration.
-          </rc.p>
-        </rc.Content>
-        <rc.Keywords>
-          <rc.Keyword value="Content Creation Inspiration" />
-          <rc.Keyword value="Motivation for Content Creators" />
-          <rc.Keyword value="Quality Content Strategies" />
-          <rc.Keyword value="Creative Community Engagement" />
-          {/* Use a normal single quote here, because users are not likely to type out the proper apostrophe character */}
-          <rc.Keyword value="Overcoming Creator's Block" />
-          <rc.Keyword value="Creator Economy Malaysia" />
-        </rc.Keywords>
+        <rc.h2>How Do I Make Good Quality Content?</rc.h2>
+        <rc.p>
+          Creating good quality content involves a combination of factors that
+          contribute to its overall appeal and value. Here are some key
+          strategies to consider when striving to produce high-quality content:
+        </rc.p>
+        <rc.ul>
+          <rc.li>
+            <b>Understanding your audience.</b> One of the fundamental aspects
+            of creating good quality content is to understand your audience.
+            Conduct thorough research to determine the demographics,
+            preferences, and interests of your target audience. This allows you
+            to tailor your content so that it resonates with them, at once
+            increasing engagement and relevance.
+          </rc.li>
+          <rc.li>
+            <b>Consistency is key.</b> Consistently creating and publishing
+            content is essential for maintaining a strong presence and
+            connection with your audience. Whether it’s a weekly blog post,
+            daily social media updates, or regular video uploads, maintaining a
+            consistent schedule helps build anticipation and reliability among
+            your audience.
+          </rc.li>
+          <rc.li>
+            <b>Quality over quantity.</b> While consistency is important, it’s
+            equally crucial to prioritise quality over quantity. Focus on
+            producing content that is well-researched, valuable, and relevant to
+            your audience’s needs and interests. Strive for originality and
+            uniqueness in your content to stand out in a crowded digital
+            landscape.
+          </rc.li>
+          <rc.li>
+            <b>Use engaging visual elements.</b> Incorporating engaging visuals
+            such as images, infographics, and videos can significantly enhance
+            the overall appeal of your content. Visual elements not only capture
+            attention, but also help in conveying your messages effectively.
+          </rc.li>
+          <rc.li>
+            <b>Storytelling and emotion.</b> Effective storytelling can evoke
+            powerful emotions and create a lasting impact on your audience.
+            Consider incorporating compelling narratives, personal anecdotes,
+            and emotional appeal in your content to establish a deeper
+            connection with your audience.
+          </rc.li>
+          <rc.li>
+            <b>Continuous improvement.</b> Always seek opportunities for
+            improvement. Stay updated with the latest trends, gather feedback
+            from your audience, and be open to learn and evolve as a content
+            creator. Embracing a mindset of continuous improvement ensures that
+            your content remains relevant and compelling in an ever-changing
+            digital landscape.
+          </rc.li>
+        </rc.ul>
+        <rc.p>
+          While it is the goal to constantly create content, it is also crucial
+          to take breaks when needed as burnout can really impact your
+          creativity. Finding a balance between work and personal life is
+          essential for long-term motivation and inspiration.
+        </rc.p>
       </rc.Page>
     </>
   );
