+++ src/app/resources/page.jsx
@@ -1,18 +1,22 @@
-"use client";
+import { makeCanonicalUrl } from "@/app/components/misc";
 import { PageMeta } from "@/app/components/PageMeta";
 import Header from "@/app/components/Header";
 import resourcesData from "@/app/contants/resources.json";
 import News from "@/app/components/News";
 import LastSection from "@/app/sections/brandsection/LastSection";
 
+export const metadata = {
+  title: "Resources",
+  description: "Discover invaluable insights and resources at Cult Creative’s blog. Dive into inspiration, expert tips, and stay ahead of industry trends with our creator-friendly content hub.",
+  alternates: {
+    canonical: makeCanonicalUrl("/resources"),
+  },
+};
+
 const Resources = () => {
   return (
     <>
-      <PageMeta
-        title="Resources"
-        desc="Discover invaluable insights and resources at Cult Creative’s blog. Dive into inspiration, expert tips, and stay ahead of industry trends with our creator-friendly content hub."
-        canonicalPath="/resources"
-      />
+      <PageMeta />
       <main className="flex min-h-screen flex-col bg-[#F4F4F4] text-white">
         <Header />
         <div className="text-center py-10 bg-[#8A5AFE]">
