+++ src/app/resources/turn-your-content-creation-passion-into-profit/page.jsx
@@ -1,169 +1,175 @@
-"use client";
-import { PageMeta } from "@/app/components/PageMeta";
 import resources from "@/app/contants/resources.json";
+import { makeCanonicalUrl } from "@/app/components/misc";
+import { PageMeta } from "@/app/components/PageMeta";
 import * as rc from "@/app/components/resource-content";
 
-const Resource3 = () => {
-  const title = "Turn Your Content Creation Passion into Profit";
+const title = "Turn Your Content Creation Passion into Profit";
 
+const keywords = [
+  "Content creation career",
+  "Content creator tips",
+  "Building an online presence",
+  "Social media engagement",
+  "Monetizing content creation",
+  "Creator economy in Malaysia",
+];
+
+export const metadata = {
+  title: title,
+  description: "Big brands today partner with creators for marketing. Learn about monetization strategies and how to engage authentically with your audience. And beware of competition and burnout.",
+  keywords: keywords,
+  openGraph: {
+    images: [{ url: resources[2].image }],
+  },
+  alternates: {
+    canonical: makeCanonicalUrl("/resources/turn-your-content-creation-passion-into-profit"),
+  },
+};
+
+const Resource3 = () => {
   return (
     <>
-      <PageMeta
-        title={title}
-        desc="Big brands today partner with creators for marketing. Learn about monetization strategies and how to engage authentically with your audience. And beware of competition and burnout."
-        canonicalPath="/resources/turn-your-content-creation-passion-into-profit"
-        ogImage={resources[2].image}
-      />
-      <rc.Page>
-        <rc.Content>
-          <rc.h1>{title}</rc.h1>
+      <PageMeta />
+      <rc.Page keywords={keywords}>
+        <rc.h1>{title}</rc.h1>
 
-          <rc.h2>How to Turn Content Creation into a Career</rc.h2>
-          <rc.p>
-            Content creation has evolved from mere hobby to an actual money-making
-            endeavour as businesses and platforms begin to leverage this trade as
-            part of their marketing campaigns. Major brands like Marc Jacobs,
-            Airbnb, and Halo Top have partnered with numerous content creators to
-            promote their products and experiences.
-          </rc.p>
-          <rc.p>
-            It is no surprise, then, that many people are keen to turn their
-            passion for content creation into a career. Fortunately, there are
-            numerous opportunities and resources available for individuals looking
-            to make money out of content creation.
-          </rc.p>
-          <rc.p>
-            One key aspect of doing so is to focus on building a strong online
-            presence. This involves creating a professional website or blog to
-            showcase your work, establishing a presence on social media platforms
-            relevant to your niche, and consistently producing high-quality
-            content that resonates with your target audience.
-          </rc.p>
+        <rc.h2>How to Turn Content Creation into a Career</rc.h2>
+        <rc.p>
+          Content creation has evolved from mere hobby to an actual money-making
+          endeavour as businesses and platforms begin to leverage this trade as
+          part of their marketing campaigns. Major brands like Marc Jacobs,
+          Airbnb, and Halo Top have partnered with numerous content creators to
+          promote their products and experiences.
+        </rc.p>
+        <rc.p>
+          It is no surprise, then, that many people are keen to turn their
+          passion for content creation into a career. Fortunately, there are
+          numerous opportunities and resources available for individuals looking
+          to make money out of content creation.
+        </rc.p>
+        <rc.p>
+          One key aspect of doing so is to focus on building a strong online
+          presence. This involves creating a professional website or blog to
+          showcase your work, establishing a presence on social media platforms
+          relevant to your niche, and consistently producing high-quality
+          content that resonates with your target audience.
+        </rc.p>
 
-          <rc.h2>
-            Building a Genuine Audience on Social Media through Quality Content
-          </rc.h2>
-          <rc.p>
-            Building a genuine audience on social media requires a combination of
-            high-quality content and strategic engagement.
-          </rc.p>
-          <rc.p>
-            An essential element of that is to create content that resonates with
-            your target audience. This involves understanding their interests,
-            pain points and preferences, as well as tailoring your content to meet
-            their needs.
-          </rc.p>
-          <rc.p>
-            Additionally, incorporating user-generated content (UGC) into your
-            social media strategy can also help foster a sense of community among
-            your audience. However, it is crucial that you stay true to your niche
-            and values when choosing to work with a brand. Working with a brand
-            that does not align with your values and your followers’ values might
-            turn your followers away from you.
-          </rc.p>
-          <rc.p>
-            Authenticity is key to when it comes to securing your audience’s trust
-            and loyalty.
-          </rc.p>
+        <rc.h2>
+          Building a Genuine Audience on Social Media through Quality Content
+        </rc.h2>
+        <rc.p>
+          Building a genuine audience on social media requires a combination of
+          high-quality content and strategic engagement.
+        </rc.p>
+        <rc.p>
+          An essential element of that is to create content that resonates with
+          your target audience. This involves understanding their interests,
+          pain points and preferences, as well as tailoring your content to meet
+          their needs.
+        </rc.p>
+        <rc.p>
+          Additionally, incorporating user-generated content (UGC) into your
+          social media strategy can also help foster a sense of community among
+          your audience. However, it is crucial that you stay true to your niche
+          and values when choosing to work with a brand. Working with a brand
+          that does not align with your values and your followers’ values might
+          turn your followers away from you.
+        </rc.p>
+        <rc.p>
+          Authenticity is key to when it comes to securing your audience’s trust
+          and loyalty.
+        </rc.p>
 
-          <rc.h2>Maximising Your Online Presence</rc.h2>
-          <rc.p>
-            In addition to building a diverse portfolio, maximising your online
-            presence is just as crucial for turning your content creation passion
-            into a successful career. This involves not only creating high-quality
-            content, but also engaging with your audience through social media,
-            online communities, and other digital platforms.
-          </rc.p>
-          <rc.h3>Get on that SEO train</rc.h3>
-          <rc.p>
-            To do so, you should consider leveraging search engine optimisation
-            (SEO) techniques to improve the visibility of your website, blog, and
-            videos. This will help attract more organic traffic and reach a wider
-            audience interested in your content.
-          </rc.p>
-          <rc.h3>Talk to your followers</rc.h3>
-          <rc.p>
-            Additionally, engaging with your audience through comments, direct
-            messages, and live interactions can help build a loyal and engaged
-            following, ultimately supporting your career growth as a content
-            creator.
-          </rc.p>
-          <rc.p>
-            Take a look at Jack Edwards, a BookTuber who has amassed a cult
-            following on his social platforms by interacting with his audience
-            through story reposts, comments, and providing them with{" "}
-            <rc.a href="https://youtu.be/_0RapuQ8Nkc?si=bUmk8S9rndBq8A3C">
-              personalised book recommendations
-            </rc.a>{" "}
-            on his YouTube channel.
-          </rc.p>
-          <rc.h3>Collab with other creators and brands</rc.h3>
-          <rc.p>
-            Collaborating with other content creators, industry influencers, and
-            brands can also expand your reach and provide valuable networking
-            opportunities. By strategically aligning yourself with like-minded
-            individuals and businesses, you can access new audiences and
-            potentially secure partnerships or sponsored opportunities that align
-            with your content creation goals. An emerging free platform that could
-            help bridge you to these connections is Cult Creative’s creator
-            platform. There, you can look around for the brands you’d like to work
-            with and pitch your content idea to them. If they like what they’re
-            seeing, you might be selected to work with them on a brand deal or
-            campaign!
-          </rc.p>
-          <rc.p>
-            Remember, as you continue to build your online presence and expand
-            your reach, it’s important to stay authentic and true to your creative
-            vision. Building a successful career as a content creator involves a
-            combination of consistent quality output, strategic networking, and
-            genuine engagement with your audience.
-          </rc.p>
+        <rc.h2>Maximising Your Online Presence</rc.h2>
+        <rc.p>
+          In addition to building a diverse portfolio, maximising your online
+          presence is just as crucial for turning your content creation passion
+          into a successful career. This involves not only creating high-quality
+          content, but also engaging with your audience through social media,
+          online communities, and other digital platforms.
+        </rc.p>
+        <rc.h3>Get on that SEO train</rc.h3>
+        <rc.p>
+          To do so, you should consider leveraging search engine optimisation
+          (SEO) techniques to improve the visibility of your website, blog, and
+          videos. This will help attract more organic traffic and reach a wider
+          audience interested in your content.
+        </rc.p>
+        <rc.h3>Talk to your followers</rc.h3>
+        <rc.p>
+          Additionally, engaging with your audience through comments, direct
+          messages, and live interactions can help build a loyal and engaged
+          following, ultimately supporting your career growth as a content
+          creator.
+        </rc.p>
+        <rc.p>
+          Take a look at Jack Edwards, a BookTuber who has amassed a cult
+          following on his social platforms by interacting with his audience
+          through story reposts, comments, and providing them with{" "}
+          <rc.a href="https://youtu.be/_0RapuQ8Nkc?si=bUmk8S9rndBq8A3C">
+            personalised book recommendations
+          </rc.a>{" "}
+          on his YouTube channel.
+        </rc.p>
+        <rc.h3>Collab with other creators and brands</rc.h3>
+        <rc.p>
+          Collaborating with other content creators, industry influencers, and
+          brands can also expand your reach and provide valuable networking
+          opportunities. By strategically aligning yourself with like-minded
+          individuals and businesses, you can access new audiences and
+          potentially secure partnerships or sponsored opportunities that align
+          with your content creation goals. An emerging free platform that could
+          help bridge you to these connections is Cult Creative’s creator
+          platform. There, you can look around for the brands you’d like to work
+          with and pitch your content idea to them. If they like what they’re
+          seeing, you might be selected to work with them on a brand deal or
+          campaign!
+        </rc.p>
+        <rc.p>
+          Remember, as you continue to build your online presence and expand
+          your reach, it’s important to stay authentic and true to your creative
+          vision. Building a successful career as a content creator involves a
+          combination of consistent quality output, strategic networking, and
+          genuine engagement with your audience.
+        </rc.p>
 
-          <rc.h2>How Do You Transition from Passion to Profit?</rc.h2>
-          <rc.p>
-            Making profit off content creation requires a strategic approach and
-            careful planning. Start by identifying your target audience and niche.
-            Then work on developing a clear understanding of your unique value
-            proposition and how you can monetise your content.
-          </rc.p>
-          <rc.p>
-            There are several ways you can monetise your content. This includes
-            sponsored collaborations, affiliate marketing, selling digital
-            products or services, and creating paid memberships or subscriptions.
-          </rc.p>
-          <rc.p>
-            While many people aspire to turn their content creation passion into a
-            career, it’s important to consider the potential challenges and
-            drawbacks associated with pursuing this path. One key aspect to
-            consider is the highly competitive nature of the industry. With the
-            increasing number of individuals seeking to establish themselves as
-            content creators, standing out and gaining recognition can be
-            incredibly challenging.
-          </rc.p>
-          <rc.p>
-            Furthermore, building a sustainable income solely from content
-            creation can be unpredictable and inconsistent, especially in the
-            early stages of your career. Securing regular paid opportunities and
-            building a loyal client base can take time and persistence. But there
-            are various platforms out there that can help you with that. Cult
-            Creative’s creator platform is one of them.
-          </rc.p>
-          <rc.p>
-            Additionally, the pressure to constantly produce high-quality content
-            to maintain engagement and relevance can lead to burnout and creative
-            fatigue. However, with dedication, perseverance, and a strategic
-            approach, it is possible to to turn your content creation passion into
-            a career.
-          </rc.p>
-        </rc.Content>
-        <rc.Keywords>
-          <rc.Keyword value="Content Creation Career" />
-          <rc.Keyword value="Content Creator Tips" />
-          <rc.Keyword value="Building an Online Presence" />
-          <rc.Keyword value="Social Media Engagement" />
-          <rc.Keyword value="Monetizing Content Creation" />
-          <rc.Keyword value="Creator Economy Malaysia" />
-        </rc.Keywords>
+        <rc.h2>How Do You Transition from Passion to Profit?</rc.h2>
+        <rc.p>
+          Making profit off content creation requires a strategic approach and
+          careful planning. Start by identifying your target audience and niche.
+          Then work on developing a clear understanding of your unique value
+          proposition and how you can monetise your content.
+        </rc.p>
+        <rc.p>
+          There are several ways you can monetise your content. This includes
+          sponsored collaborations, affiliate marketing, selling digital
+          products or services, and creating paid memberships or subscriptions.
+        </rc.p>
+        <rc.p>
+          While many people aspire to turn their content creation passion into a
+          career, it’s important to consider the potential challenges and
+          drawbacks associated with pursuing this path. One key aspect to
+          consider is the highly competitive nature of the industry. With the
+          increasing number of individuals seeking to establish themselves as
+          content creators, standing out and gaining recognition can be
+          incredibly challenging.
+        </rc.p>
+        <rc.p>
+          Furthermore, building a sustainable income solely from content
+          creation can be unpredictable and inconsistent, especially in the
+          early stages of your career. Securing regular paid opportunities and
+          building a loyal client base can take time and persistence. But there
+          are various platforms out there that can help you with that. Cult
+          Creative’s creator platform is one of them.
+        </rc.p>
+        <rc.p>
+          Additionally, the pressure to constantly produce high-quality content
+          to maintain engagement and relevance can lead to burnout and creative
+          fatigue. However, with dedication, perseverance, and a strategic
+          approach, it is possible to to turn your content creation passion into
+          a career.
+        </rc.p>
       </rc.Page>
     </>
   );
