+++ src/app/resources/unlocking-the-power-of-brand-deals-a-guide-for-content-creators/page.jsx
@@ -1,119 +1,125 @@
-"use client";
-import { PageMeta } from "@/app/components/PageMeta";
 import resources from "@/app/contants/resources.json";
+import { makeCanonicalUrl } from "@/app/components/misc";
+import { PageMeta } from "@/app/components/PageMeta";
 import * as rc from "@/app/components/resource-content";
 
-const Resource1 = () => {
-  const title = "Unlocking the Power of Brand Deals: A Guide for Content Creators";
+const title = "Unlocking the Power of Brand Deals: A Guide for Content Creators";
 
+const keywords = [
+  "Creator economy in Malaysia",
+  "Brand deals for creators",
+  "Content creator tips",
+  "Authentic brand collaborations",
+  "Negotiating brand contracts",
+];
+
+export const metadata = {
+  title: title,
+  description: "A guide to securing brand deals with Cult Creative to simplify the process of connecting with brands, letting creators grow their careers authentically and lucratively.",
+  keywords: keywords,
+  openGraph: {
+    images: [{ url: resources[0].image }],
+  },
+  alternates: {
+    canonical: makeCanonicalUrl("/resources/unlocking-the-power-of-brand-deals-a-guide-for-content-creators"),
+  },
+};
+
+const Resource1 = () => {
   return (
     <>
-      <PageMeta
-        title={title}
-        desc="A guide to securing brand deals with Cult Creative to simplify the process of connecting with brands, letting creators grow their careers authentically and lucratively."
-        canonicalPath="/resources/unlocking-the-power-of-brand-deals-a-guide-for-content-creators"
-        ogImage={resources[0].image}
-      />
-      <rc.Page>
-        <rc.Content>
-          <rc.h1>{title}</rc.h1>
-          <rc.p>
-            Content creators have become a pivotal part of social media marketing
-            and entertainment. With the creator economy booming, they now have a
-            higher chance at earning money and popularity through brand deals and
-            partnership, bringing both the creators and businesses more
-            visibility, credibility, and financial success.
-          </rc.p>
-          <rc.p>
-            But navigating the world of content creation isn’t always easy and
-            securing brand deals is even harder, so here’s a guide on how you can
-            land your next brand deal.
-          </rc.p>
+      <PageMeta />
+      <rc.Page keywords={keywords}>
+        <rc.h1>{title}</rc.h1>
+        <rc.p>
+          Content creators have become a pivotal part of social media marketing
+          and entertainment. With the creator economy booming, they now have a
+          higher chance at earning money and popularity through brand deals and
+          partnership, bringing both the creators and businesses more
+          visibility, credibility, and financial success.
+        </rc.p>
+        <rc.p>
+          But navigating the world of content creation isn’t always easy and
+          securing brand deals is even harder, so here’s a guide on how you can
+          land your next brand deal.
+        </rc.p>
 
-          <rc.h2>Taking on Brand Deals</rc.h2>
-          <rc.p>
-            There’s nothing sweeter than earning your first brand deal. You’ll get
-            to monetise your content (yay, more money!), collaborate with your
-            dream brands, and even reach larger audiences. By following these
-            tips, you can make the most out of your future brand deals.
-          </rc.p>
-          <rc.ol>
-            <rc.li>
-              <b>Authenticity is key.</b> Your audience trusts you for your
-              authenticity, so it’s important to collaborate with brands that
-              align with your values and that resonate with your audience.
-            </rc.li>
-            <rc.li>
-              <b>Transparency in disclosure.</b> No one likes a liar and no one
-              likes to get sued either. Always be transparent about your content
-              by adhering to relevant advertising guidelines and clearly
-              disclosing when a content is sponsored.
-            </rc.li>
-            <rc.li>
-              <b>Quality over quantity.</b> Instead of pursuing numerous brand
-              deals, prioritise the ones that will allow you to produce engaging
-              and authentic content that adds value to both the brand and your
-              audience.
-            </rc.li>
-            <rc.li>
-              <b>Negotiation and contracts.</b> Familiarise yourself with
-              negotiation tactics and remember to always navigate contracts with
-              caution. Make sure the terms of the contract aligns with your brand
-              and audience, and don’t hesitate to seek legal guidance when
-              necessary.
-            </rc.li>
-            <rc.li>
-              <b>Consistent brand alignment.</b> Maintain a consistent brand image
-              and messaging across all your brand deals. This ensures cohesive and
-              authentic representation of your brand and values.
-            </rc.li>
-          </rc.ol>
-          <rc.p>
-            Stick to this guide when approaching brand deals, so you can establish
-            yourself as a credible and sought-after content creator whilst also
-            maintaining the trust of your audience.
-          </rc.p>
+        <rc.h2>Taking on Brand Deals</rc.h2>
+        <rc.p>
+          There’s nothing sweeter than earning your first brand deal. You’ll get
+          to monetise your content (yay, more money!), collaborate with your
+          dream brands, and even reach larger audiences. By following these
+          tips, you can make the most out of your future brand deals.
+        </rc.p>
+        <rc.ol>
+          <rc.li>
+            <b>Authenticity is key.</b> Your audience trusts you for your
+            authenticity, so it’s important to collaborate with brands that
+            align with your values and that resonate with your audience.
+          </rc.li>
+          <rc.li>
+            <b>Transparency in disclosure.</b> No one likes a liar and no one
+            likes to get sued either. Always be transparent about your content
+            by adhering to relevant advertising guidelines and clearly
+            disclosing when a content is sponsored.
+          </rc.li>
+          <rc.li>
+            <b>Quality over quantity.</b> Instead of pursuing numerous brand
+            deals, prioritise the ones that will allow you to produce engaging
+            and authentic content that adds value to both the brand and your
+            audience.
+          </rc.li>
+          <rc.li>
+            <b>Negotiation and contracts.</b> Familiarise yourself with
+            negotiation tactics and remember to always navigate contracts with
+            caution. Make sure the terms of the contract aligns with your brand
+            and audience, and don’t hesitate to seek legal guidance when
+            necessary.
+          </rc.li>
+          <rc.li>
+            <b>Consistent brand alignment.</b> Maintain a consistent brand image
+            and messaging across all your brand deals. This ensures cohesive and
+            authentic representation of your brand and values.
+          </rc.li>
+        </rc.ol>
+        <rc.p>
+          Stick to this guide when approaching brand deals, so you can establish
+          yourself as a credible and sought-after content creator whilst also
+          maintaining the trust of your audience.
+        </rc.p>
 
-          <rc.h2>How Do You Get Brand Deals on Cult Creative?</rc.h2>
-          <rc.p>
-            Scoring your first (or second or third) brand deal just got easier
-            thanks to{" "}
-            <rc.a href="https://www.instagram.com/cultcreativeasia/">
-              Cult Creative’s creator platform
-            </rc.a>.
-            Whether you’re into fashion and beauty, tech and finance, or maybe
-            you’re an explorer who’s ready to seize any opportunity that comes
-            your way, we’ll connect you to amazing brands across Southeast Asia
-            when you become a CC creator.
-          </rc.p>
-          <rc.p>Did I mention it’s free to use? Yup, it is!</rc.p>
-          <rc.p>
-            Simply fill out{" "}
-            <rc.a href="https://airtable.com/appMK1DimSupLgFeU/shrc6kh4QsG1N0geq">
-              this form
-            </rc.a>{" "}
-            and you’re on your way to securing your next brand deal.
-          </rc.p>
+        <rc.h2>How Do You Get Brand Deals on Cult Creative?</rc.h2>
+        <rc.p>
+          Scoring your first (or second or third) brand deal just got easier
+          thanks to{" "}
+          <rc.a href="https://www.instagram.com/cultcreativeasia/">
+            Cult Creative’s creator platform
+          </rc.a>.
+          Whether you’re into fashion and beauty, tech and finance, or maybe
+          you’re an explorer who’s ready to seize any opportunity that comes
+          your way, we’ll connect you to amazing brands across Southeast Asia
+          when you become a CC creator.
+        </rc.p>
+        <rc.p>Did I mention it’s free to use? Yup, it is!</rc.p>
+        <rc.p>
+          Simply fill out{" "}
+          <rc.a href="https://airtable.com/appMK1DimSupLgFeU/shrc6kh4QsG1N0geq">
+            this form
+          </rc.a>{" "}
+          and you’re on your way to securing your next brand deal.
+        </rc.p>
 
-          <rc.h2>Conclusion</rc.h2>
-          <rc.p>
-            It can be both exciting and complex to work with brands. It is the
-            path to elevating your career and creative endeavours.
-          </rc.p>
-          <rc.p>
-            With this guide, you’ll be well-equipped to forge successful
-            partnerships with brands whilst also retaining the trust and loyalty
-            of your audience. Remember to always be authentic, transparent, and
-            genuine about your content no matter who you’re creating it for.
-          </rc.p>
-        </rc.Content>
-        <rc.Keywords>
-          <rc.Keyword value="Creator Economy Malaysia" />
-          <rc.Keyword value="Brand Deals for Creators" />
-          <rc.Keyword value="Content Creator Tips" />
-          <rc.Keyword value="Authentic Brand Collaborations" />
-          <rc.Keyword value="Negotiating Brand Contracts" />
-        </rc.Keywords>
+        <rc.h2>Conclusion</rc.h2>
+        <rc.p>
+          It can be both exciting and complex to work with brands. It is the
+          path to elevating your career and creative endeavours.
+        </rc.p>
+        <rc.p>
+          With this guide, you’ll be well-equipped to forge successful
+          partnerships with brands whilst also retaining the trust and loyalty
+          of your audience. Remember to always be authentic, transparent, and
+          genuine about your content no matter who you’re creating it for.
+        </rc.p>
       </rc.Page>
     </>
   );
