+++ src/app/resources/why-and-how-does-creator-marketing-work/page.jsx
@@ -1,95 +1,101 @@
-"use client";
-import { PageMeta } from "@/app/components/PageMeta";
 import resources from "@/app/contants/resources.json";
+import { makeCanonicalUrl } from "@/app/components/misc";
+import { PageMeta } from "@/app/components/PageMeta";
 import * as rc from "@/app/components/resource-content";
 
-const Resource6 = () => {
-  const title = "Why and How Does Creator Marketing Work?";
+const title = "Why and How Does Creator Marketing Work?";
 
+const keywords = [
+  "Creator marketing benefits",
+  "Influencer marketing advantages",
+  "Building brand trust with influencers",
+  "Effective UGC strategies",
+  "Simplified UGC campaign management",
+];
+
+export const metadata = {
+  title: title,
+  description: "Creator marketing thrives on authenticity. Brands can easily initiate UGC campaigns through platforms like Cult Creative, which connects brands with suitable creators hassle-free.",
+  keywords: keywords,
+  openGraph: {
+    images: [{ url: resources[5].image }],
+  },
+  alternates: {
+    canonical: makeCanonicalUrl("/resources/why-and-how-does-creator-marketing-work"),
+  },
+};
+
+const Resource6 = () => {
   return (
     <>
-      <PageMeta
-        title={title}
-        desc="Creator marketing thrives on authenticity. Brands can easily initiate UGC campaigns through platforms like Cult Creative, which connects brands with suitable creators hassle-free."
-        canonicalPath="/resources/why-and-how-does-creator-marketing-work"
-        ogImage={resources[5].image}
-      />
-      <rc.Page>
-        <rc.Content>
-          <rc.h1>{title}</rc.h1>
-          <rc.p>
-            Here’s the short answer: people no longer want brands to simply sell to them.
-          </rc.p>
-          <rc.p>
-            By leveraging the influence and credibility of content creators or influencers to promote products or brands to their audience, brands have the unique opportunity to reach and engage with a highly targeted and receptive customer base. Why? Because the audience already has a connection with the creator, making their recommendations and endorsements more powerful and persuasive.
-          </rc.p>
-          <rc.p>
-            As consumers increasingly seek genuine and relatable experiences, influencer marketing allows brands to partner with creators who produce content that resonates with their target audience.
-          </rc.p>
+      <PageMeta />
+      <rc.Page keywords={keywords}>
+        <rc.h1>{title}</rc.h1>
+        <rc.p>
+          Here’s the short answer: people no longer want brands to simply sell to them.
+        </rc.p>
+        <rc.p>
+          By leveraging the influence and credibility of content creators or influencers to promote products or brands to their audience, brands have the unique opportunity to reach and engage with a highly targeted and receptive customer base. Why? Because the audience already has a connection with the creator, making their recommendations and endorsements more powerful and persuasive.
+        </rc.p>
+        <rc.p>
+          As consumers increasingly seek genuine and relatable experiences, influencer marketing allows brands to partner with creators who produce content that resonates with their target audience.
+        </rc.p>
 
-          <rc.h2>Key Reasons That Creator Marketing Works</rc.h2>
-          <rc.p>
-            It works because it taps into the authenticity and relatability of content creators.
-          </rc.p>
-          <rc.p>
-            UGC (user-generated content) type of content creates a seamless and organic experience for the audience, increasing consideration and purchase intent.
-          </rc.p>
-          <rc.p>
-            Here’s more proof of how it works:
-          </rc.p>
-          <rc.ol>
-            <rc.li>
-              <b>Trust and credibility.</b> Content creators have built trust and credibility with their audience, making their recommendations more reliable and influential.
-            </rc.li>
-            <rc.li>
-              <b>Authenticity.</b> Creator marketing allows brands to connect with consumers in a more authentic way, as content creators are seen as real people rather than traditional advertisements.
-            </rc.li>
-            <rc.li>
-              <b>Audience engagement.</b> Influencers have a direct line of communication with their audience and can engage with them in a more interactive and personalised manner, leading to increased engagement and loyalty.
-            </rc.li>
-            <rc.li>
-              <b>Targeted reach.</b> Content creators often have a niche or specific audience, allowing brands to reach and engage with a highly targeted group of consumers.
-            </rc.li>
-            <rc.li>
-              <b>Expertise and authority.</b> Influencers are often experts or knowledgeable in their niche, which adds credibility to the brand they endorse and increases consumer confidence in the product or service.
-            </rc.li>
-          </rc.ol>
+        <rc.h2>Key Reasons That Creator Marketing Works</rc.h2>
+        <rc.p>
+          It works because it taps into the authenticity and relatability of content creators.
+        </rc.p>
+        <rc.p>
+          UGC (user-generated content) type of content creates a seamless and organic experience for the audience, increasing consideration and purchase intent.
+        </rc.p>
+        <rc.p>
+          Here’s more proof of how it works:
+        </rc.p>
+        <rc.ol>
+          <rc.li>
+            <b>Trust and credibility.</b> Content creators have built trust and credibility with their audience, making their recommendations more reliable and influential.
+          </rc.li>
+          <rc.li>
+            <b>Authenticity.</b> Creator marketing allows brands to connect with consumers in a more authentic way, as content creators are seen as real people rather than traditional advertisements.
+          </rc.li>
+          <rc.li>
+            <b>Audience engagement.</b> Influencers have a direct line of communication with their audience and can engage with them in a more interactive and personalised manner, leading to increased engagement and loyalty.
+          </rc.li>
+          <rc.li>
+            <b>Targeted reach.</b> Content creators often have a niche or specific audience, allowing brands to reach and engage with a highly targeted group of consumers.
+          </rc.li>
+          <rc.li>
+            <b>Expertise and authority.</b> Influencers are often experts or knowledgeable in their niche, which adds credibility to the brand they endorse and increases consumer confidence in the product or service.
+          </rc.li>
+        </rc.ol>
 
-          <rc.h2>Get Started with Your Brand UGC Campaigns</rc.h2>
-          <rc.p>
-            To get started with creator marketing and run successful paid UGC campaigns, start with{" "}
-            <rc.a href="https://www.cultcreative.asia/hire-a-content-creation-team-with-cult-creative">
-              Cult Creative
-            </rc.a>.
-          </rc.p>
-          <rc.p>
-            We’ve set it up so it’s easy for you as a brand to ditch the hassle that’s related to UGC and influencer marketing.
-          </rc.p>
-          <rc.p>
-            We’ll tap into our 10,000-strong creator base to find one that suits you best.
-          </rc.p>
-          <rc.ul>
-            <rc.li>
-              Simply fill out a brief form and indicate the level of budget—and how many UGC videos you need.
-            </rc.li>
-            <rc.li>
-              A Customer Success Manager will be in touch with you within a day to confirm the details of the brief.
-            </rc.li>
-            <rc.li>
-              Discuss, lock it in and that’s it! Project commences and all you need to do is approve creators and content, while it gets posted and runs for you.
-            </rc.li>
-          </rc.ul>
-          <rc.p>
-            It’s really that simple!
-          </rc.p>
-        </rc.Content>
-        <rc.Keywords>
-          <rc.Keyword value="Creator Marketing Benefits" />
-          <rc.Keyword value="Influencer Marketing Advantages" />
-          <rc.Keyword value="Building Brand Trust with Influencers" />
-          <rc.Keyword value="Effective UGC Strategies" />
-          <rc.Keyword value="Simplified UGC Campaign Management" />
-        </rc.Keywords>
+        <rc.h2>Get Started with Your Brand UGC Campaigns</rc.h2>
+        <rc.p>
+          To get started with creator marketing and run successful paid UGC campaigns, start with{" "}
+          <rc.a href="https://www.cultcreative.asia/hire-a-content-creation-team-with-cult-creative">
+            Cult Creative
+          </rc.a>.
+        </rc.p>
+        <rc.p>
+          We’ve set it up so it’s easy for you as a brand to ditch the hassle that’s related to UGC and influencer marketing.
+        </rc.p>
+        <rc.p>
+          We’ll tap into our 10,000-strong creator base to find one that suits you best.
+        </rc.p>
+        <rc.ul>
+          <rc.li>
+            Simply fill out a brief form and indicate the level of budget—and how many UGC videos you need.
+          </rc.li>
+          <rc.li>
+            A Customer Success Manager will be in touch with you within a day to confirm the details of the brief.
+          </rc.li>
+          <rc.li>
+            Discuss, lock it in and that’s it! Project commences and all you need to do is approve creators and content, while it gets posted and runs for you.
+          </rc.li>
+        </rc.ul>
+        <rc.p>
+          It’s really that simple!
+        </rc.p>
       </rc.Page>
     </>
   );
