+++ src/app/sections/brandsection/FourthSection.jsx
@@ -1,3 +1,4 @@
+"use client";
 import { motion } from "framer-motion";
 import Link from "next/link";
