+++ src/app/sections/homeSection/VideoBig.jsx
@@ -38,7 +38,7 @@ const VideoBig = () => {
         autoPlay
         loop
         muted
-        webkit-playsInline
+        webkit-playsinline="true"
         playsInline
       >
         <source type="video/mp4" />
