+++ src/app/sections/homeSection/VideoSmall.jsx
@@ -1,5 +1,3 @@
-import React from "react";
-
 const VideoSmall = ({ src }) => {
   return (
     <div>
